Did a number of changes.  Mostly to support the non-WEB generation of reports.
added routines to send email:
	+ email_senddoc()
	+ email_wfiles()
	+ email_simple()
added routines to output data as text-table (eg.psql output) & html-table:
	+ data_to_htmltable()
	+ data_to_texttable(): 	| column A |   column B |
				-------------------------
				| JeremyE  |  3/14/2012 |
changed the way the converter routines work: broke each into a pair of routines:
	data_to_*() : convert data to specific format
	*_response(): produce an http response appropriate for this type
the main purpose for this change was to allow the same routines to be used for
generating formatted documents for email sending.  Email needs docs, NOT http responses,
so the logic needed to be pulled apart.
Also the data_to_*() methods were extended to return the document, AND the content-type,
and mime-type to use for this format.
	+ data_to_gjson
	+ gjson_response
	+ data_to_json
	+ json_response
	+ data_to_csv
	+ csv_response
Added routines to store output files:
	* store doc to s3
	* store doc to local
	* tinyurl  -- doesn't work yet!
Added routines to allow caching datasets in mongo:
	* list mongo fixtures
	* load mongo fixture
	* store as mongo